+++ CHANGELOG.md
@@ -19,7 +19,7 @@ AdminNeo 5.6.0
 ### Bugfixes
 
 - Ignore invalid X-Forwarded-Prefix (by @vrana, GHSA-5fw2-qrc6-p4fw)
-- Support disabled set_time_limit() (by @vrana)
+- Support disabled set_time_limit() and ini_set() (by @vrana)
 - MariaDB: Allow setting password with password validation plugin (by @vrana)
 - SQLite: Preserve STRICT and WITHOUT ROWID in alter table (by @vrana)
 - SQLite: Do not treat INTEGER PRIMARY KEY without AUTOINCREMENT as auto increment (by @vrana)